correct ProductRepository and UserController

+++ homework-spring-3/src/main/java/ru/learning/java/spring/repository/ProductRepository.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Repository;
 import ru.learning.java.spring.model.Product;
 import ru.learning.java.spring.model.ProductType;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
@@ -20,5 +21,5 @@ public interface ProductRepository extends JpaRepository<Product, Long> {
   List<Product> findByUserIdAndProductType(@Param("userId") Long userId, @Param("productType") ProductType productType);
 
   @Query("SELECT p FROM Product p WHERE p.balance >= :minBalance")
-  List<Product> findByBalanceGreaterThanEqual(@Param("minBalance") Double minBalance);
+  List<Product> findByBalanceGreaterThanEqual(@Param("minBalance") BigDecimal minBalance);
 }
